**- What I did**

Since commit https://github.com/moby/moby/commit/933fcc981418bd6994abd76701a2833789e143f5 (Re-remove the SetKey OCI prestart hook), the network namespace will be set up before endpoints are created in most cases (apart from build containers).

So, when possible, create the veth with one end in that netns to save moving it in later.

On my Mac M2 host, that reduces container startup time by about 20ms for each bridge network it's connected to.

**- How I did it**

Added methods for a driver to ask for the network namespace, and to announce that it's created the interface in the container's netns ... if anything goes wrong, fall back to the old behaviour, creating the veth in the host's namespace.

Tell the `osl` code when it doesn't need to move the interface into the namespace.

No change to the `srcName` property, the name the interface would have been created with in the host namespace. When created in the container's namespace, the interface has that same name (it has to be renamed, as it was before).

**- How to verify it**

New and existing tests.

Before - `moveLink` in this run took 26ms of the total 94ms ...

<img width="1670" alt="before" src="https://github.com/user-attachments/assets/c69faa90-5e57-4f6e-b656-26911e6e5ca8" />

After - `moveLink` isn't called, reducing the total time to 68ms ...

<img width="1674" alt="after" src="https://github.com/user-attachments/assets/efd2a9e9-3a51-44db-92fc-a191d8e10008" />

**- Description for the changelog**
```markdown changelog
Faster connection to bridge networks, in most cases.
```