@robmry 2025-01-23
The second of the two commits I just added is to make sure `waitForIfUpped` is looking at the interface in the container's netns ... originally, `moveLink` returned the netns handle it should use, but it's not called in the new flow. So, things appeared to work when the host's netns had an interface with the same ifIndex as the container's interface, but a 5s timeout was hit when the host had no such interface. (And the first commit would have made that easier to debug.)